Refactor new unit tests. Update CHANGELOG & README. Update __template.py. Clean up formatting/comments.
