+++ unit_tests/generate_branch_list_csv_test.py
@@ -1,18 +1,12 @@
 #!/usr/bin/env python3
 
-import inspect
 import os
-import sys
-
 import json
-import warnings
 import unittest
+import pytest
 
-sys.path.append('/foss_fim/unit_tests/')
 from unit_tests_utils import FIM_unit_test_helpers as ut_helpers
 
-# importing python folders in other directories
-sys.path.append('/foss_fim/src/')   # *** update your folder path here if required ***
 import generate_branch_list_csv as src
 
 class test_generate_branch_list_csv(unittest.TestCase):
@@ -23,7 +17,6 @@ class test_generate_branch_list_csv(unittest.TestCase):
     @classmethod
     def setUpClass(self):
 
-        warnings.simplefilter('ignore')
         params_file_path = ut_helpers.get_params_filename(__file__)
         with open(params_file_path) as params_file:
             self.params = json.load(params_file)
@@ -34,7 +27,8 @@ class test_generate_branch_list_csv(unittest.TestCase):
         params = self.params["valid_data_add_branch_zero"].copy()
         if (os.path.exists(params["output_branch_csv"])):
             os.remove(params["output_branch_csv"])
-        
+
+# Test Cases:      
 
     def test_generate_branch_list_csv_valid_data_add_branch_zero_success(self):
 
@@ -47,9 +41,6 @@ class test_generate_branch_list_csv(unittest.TestCase):
                                      branch_id = params["branch_id"],
                                      output_branch_csv = params["output_branch_csv"])
        
-       
-        print(f"Test Success: {inspect.currentframe().f_code.co_name}")
-        print("*************************************************************")
 
 
     def test_generate_branch_list_csv_valid_data_add_branch_success(self):
@@ -63,42 +54,15 @@ class test_generate_branch_list_csv(unittest.TestCase):
                                      branch_id = params["branch_id"],
                                      output_branch_csv = params["output_branch_csv"])
        
-       
-        print(f"Test Success: {inspect.currentframe().f_code.co_name}")
-        print("*************************************************************")
 
     def test_generate_branch_list_csv_invalid_bad_file_extension(self):
         
         #global params_file
         params = self.params["invalid_bad_file_extension"].copy()
 
-        # we expect this to fail. If it does fail with an exception, then this test
-        # is sucessful.
-        try:
+        # we expect this to fail. If it does fail with an exception, then this test is sucessful.
+        with pytest.raises(Exception) as e_info:
             src.generate_branch_list_csv(huc_id = params["huc_id"],
                                         branch_id = params["branch_id"],
                                         output_branch_csv = params["output_branch_csv"])
-       
-            raise AssertionError("Fail = excepted a thrown exception but did not get it but was received. Unit Test has 'failed'")
-            
-        except Exception:
-            print()
-            print(f"Test Success (failed as expected): {inspect.currentframe().f_code.co_name}")
-       
-        finally:
-            print("*************************************************************")             
 
-
-if __name__ == '__main__':
-
-    script_file_name = os.path.basename(__file__)
-
-    print("*****************************")
-    print(f"Start of {script_file_name} tests")
-    print()
-   
-    unittest.main()
-    
-    print()    
-    print(f"End of {script_file_name} tests")
-
