+++ unit_tests/generate_branch_list_params.json
@@ -1,8 +1,8 @@
 {
 	"valid_data":
 	{
-		"stream_network_dissolved": "/outputs/rob_wb_3/05030104/nwm_subset_streams_levelPaths_dissolved.gpkg",
+		"stream_network_dissolved": "/outputs/fim_unit_test_data_do_not_remove/05030104/nwm_subset_streams_levelPaths_dissolved.gpkg",
 		"branch_id_attribute": "levpa_id",
-		"output_branch_list_file": "/outputs/rob_wb_3/05030104/branch_ids.lst"
+		"output_branch_list_file": "/outputs/fim_unit_test_data_do_not_remove/05030104/branch_ids.lst"
 	}
 }
